feat: fixed output, text > byte, remove duplicate, propagate out of execscr feat: save state of hide/show code on reload
