+++ ackrep_core/util.py
@@ -1,6 +1,7 @@
 from colorama import Style, Fore
 from django.utils import timezone
 import yaml
+import subprocess
 
 import os
 from ipydex import Container
@@ -118,3 +119,39 @@ def smooth_timedelta(start_datetime, end_datetime=None):
     if secs > 0:
         timetot += " {}s".format(int(secs))
     return timetot
+
+def utf8decode(obj):
+    if hasattr(obj, "decode"):
+        return obj.decode("utf8")
+    else:
+        return obj
+
+
+def strip_decode(obj) -> str:
+
+    # get rid of some (ipython-related boilerplate bytes (ended by \x07))
+    delim = b"\x07"
+    obj = obj.split(delim)[-1]
+    return utf8decode(obj)
+
+
+def run_command(arglist, suppress_output=False, capture_output=True, **kwargs):
+    """
+    Unified handling of calling commands.
+    Automatically prints an error message if necessary.
+    """
+    res = subprocess.run(arglist, capture_output=capture_output, **kwargs)
+    res.exited = res.returncode
+    res.stdout = strip_decode(res.stdout)
+    res.stderr = strip_decode(res.stderr)
+    if res.returncode != 0 and not suppress_output:
+        msg = f"""
+        The command `{' '.join(arglist)}` exited with returncode {res.returncode}.
+
+        stdout: {res.stdout}
+
+        stderr: {res.stderr}
+        """
+        print(msg)
+
+    return res
